createUser functioning on new auth/register route
