+++ src/routes/index.ts
@@ -1,12 +1,12 @@
 import { Router } from "express";
 import apiRoutes from "./api/index";
 import authRoutes from "./authRoutes/authRoutes";
-import { authenticateToken } from "../middleware/auth";
+// import { authenticateToken } from "../middleware/auth";
 
 const router = Router();
 
 router.use("/auth", authRoutes);
-router.use("/api", authenticateToken, apiRoutes);
+router.use("/api", apiRoutes);
 
 router.use((_req, res) => {
   return res.send("404 Not Found");
